core/requirements.txt
asgiref==3.7.2
Babel==2.12.1
beautifulsoup4==4.12.2
da-vinci==0.4.0
Django==4.2.3
django-appconf==1.0.5
django-debug-toolbar==4.2.0
django-imagekit==4.1.0
django-phonenumber-field==7.1.0
django-resized==1.0.2
django-taggit==4.0.0
django-widget-tweaks==1.5.0
gunicorn==21.2.0
packaging==23.1
phonenumbers==8.13.17
pilkit==2.0
Pillow==9.5.0
psycopg-binary==3.1.10
psycopg2==2.9.6
pycodestyle==2.11.0
python-dotenv==1.0.0
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
